+++ mcp_hydrolix/auth/__init__.py
@@ -1,0 +1,26 @@
+"""Authentication package for MCP Hydrolix.
+
+This package contains authentication-related types used to define hydrolix auth
+in terms of FastMCP infrastructure
+"""
+from mcp_hydrolix.auth.credentials import (
+    HydrolixCredential,
+    ServiceAccountToken,
+    UsernamePassword,
+)
+from mcp_hydrolix.auth.mcp_providers import (
+    AccessToken,
+    ChainedAuthBackend,
+    GetParamAuthBackend,
+    HydrolixCredentialChain,
+)
+
+__all__ = [
+    "HydrolixCredential",
+    "ServiceAccountToken",
+    "UsernamePassword",
+    "AccessToken",
+    "ChainedAuthBackend",
+    "GetParamAuthBackend",
+    "HydrolixCredentialChain",
+]
